+++ dotcom-rendering/src/components/AffiliateDisclaimer.tsx
