+++ dotcom-rendering/src/components/TopMetaNews.amp.tsx
@@ -5,9 +5,9 @@ import { getAgeWarning } from '../lib/age-warning';
 import { pillarPalette_DO_NOT_USE } from '../lib/pillars';
 import { getSharingUrls } from '../lib/sharing-urls';
 import type { AMPArticleModel } from '../types/article.amp';
+import { AffiliateDisclaimer } from './AffiliateDisclaimer.amp';
 import { Branding, BrandingRegionContainer } from './Branding.amp';
 import { Byline } from './Byline.amp';
-import { Disclaimer } from './Disclaimer.amp';
 import { MainMedia } from './MainMedia.amp';
 import { SeriesLink } from './SeriesLink.amp';
 import { Standfirst } from './Standfirst.amp';
@@ -99,7 +99,7 @@ export const TopMetaNews = ({
 			<Standfirst text={articleData.standfirst} pillar={pillar} />
 
 			{!!articleData.affiliateLinksDisclaimer && (
-				<Disclaimer
+				<AffiliateDisclaimer
 					html={articleData.affiliateLinksDisclaimer}
 					pillar={pillar}
 				/>
